+++ src/types/DATATYPE.ts
@@ -1,0 +1,6 @@
+export enum DATATYPE {
+	FLOAT = "float",
+	BOOLEAN = "boolean",
+	STRING = "string",
+	INT = "int",
+}
