+++ src/types/Properties.ts
@@ -1,0 +1,4 @@
+import type { Defaults } from "./Defaults";
+export type Properties<T extends Defaults> = Partial<{
+	[K in keyof T]: T[K]["value"];
+}>;
