+++ src/utils/initValue.ts
@@ -1,0 +1,3 @@
+export function initValue<T>(userVal: T | undefined, defaultVal: T) {
+	return userVal === undefined ? defaultVal : userVal;
+}
