+++ src/utils/pipe.ts
@@ -1,0 +1,3 @@
+export const pipe = <T>(param: { value: T }, val: T) => {
+	param.value = val;
+};
