+++ src/utils/sign.ts
@@ -1,0 +1,6 @@
+export function sign(x: number) {
+	if (x === 0) {
+		return 1;
+	}
+	return Math.abs(x) / x;
+}
