+++ src/utils/tanh.ts
@@ -1,0 +1,3 @@
+export function tanh(n: number) {
+	return (Math.exp(n) - Math.exp(-n)) / (Math.exp(n) + Math.exp(-n));
+}
